add swarm.session([opts])

+++ index.js
@@ -57,6 +57,10 @@ module.exports = class Hyperswarm extends EventEmitter {
     this.explicitPeers = new Set()
     this.listening = null
 
+    this._opts = opts
+    this._root = opts._root
+    if (this._root) this._root.once('predestroy', () => this.destroy())
+
     this._discovery = new Map()
     this._timer = new RetryTimer(this._requeue.bind(this), {
       backoffs: opts.backoffs,
@@ -73,7 +77,8 @@ module.exports = class Hyperswarm extends EventEmitter {
     this._serverConnections = 0
     this._firewall = firewall
 
-    this.dht.on('network-change', this._handleNetworkChange.bind(this))
+    this._handleNetworkChange = this._handleNetworkChange.bind(this)
+    this.dht.on('network-change', this._handleNetworkChange)
   }
 
   _maybeRelayConnection (force) {
@@ -453,9 +458,24 @@ module.exports = class Hyperswarm extends EventEmitter {
     return cleared
   }
 
+  session (opts = {}) {
+    return new Hyperswarm({
+      ...this._opts,
+      seed: undefined,
+      keyPair: undefined,
+      ...opts,
+      dht: this.dht,
+      _root: this._root || this
+    })
+  }
+
   async destroy ({ force } = {}) {
     if (this.destroyed && !force) return
+
     this.destroyed = true
+    this.emit('predestroy')
+
+    this.dht.off('network-change', this._handleNetworkChange)
 
     this._timer.destroy()
 
@@ -468,7 +488,13 @@ module.exports = class Hyperswarm extends EventEmitter {
       flush.onflush(false)
     }
 
-    await this.dht.destroy({ force })
+    for (const conn of this._allConnections) {
+      conn.destroy()
+    }
+
+    if (!this._root) await this.dht.destroy({ force })
+
+    this.emit('close')
   }
 
   async suspend () {
